pyproject.toml
[tool.poetry]
name = "enerator"
version = "0.1.0"
description = ""
authors = ["Jonathan Bowman <git@bowmanjd.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/bowmanjd/enerator"
keywords = ["blog", "website", "static site generator", "static sites", "ssg"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Text Processing :: Markup :: HTML",
]

[tool.poetry.dependencies]
python = "^3.6"
python-frontmatter = "^1.0.0"
Jinja2 = "^2.11.3"
typer = "^0.3.2"
cmarkgfm = "^0.5.3"
Pygments = "^2.8.1"

mypy = { version = "^0.812", optional = true }
darglint = { version = "^1.7.0", optional = true }
flake8 = { version = "^3.9.0", optional = true }
flake8-bandit = { version = "^2.1.2", optional = true }
flake8-bugbear = { version = "^21.3.2", optional = true }
flake8-builtins = { version = "^1.5.3", optional = true }
flake8-comprehensions = { version = "^3.3.1", optional = true }
flake8-docstrings = { version = "^1.5.0", optional = true }
flake8-eradicate = { version = "^1.0.0", optional = true }
flake8-isort = { version = "^4.0.0", optional = true }
flake8-pytest-style = { version = "^1.4.0", optional = true }
flake8-spellcheck = { version = "^0.24.0", optional = true }
pep8-naming = { version = "^0.11.1", optional = true }
pynvim = { version = "^0.4.3", optional = true }
black = {version = "^20.8b1", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = {extras = ["toml"], version = "^5.5"}

[tool.poetry.extras]
lint = [
  "mypy",
  "darglint",
  "flake8",
  "flake8-bandit",
  "flake8-bugbear",
  "flake8-builtins",
  "flake8-comprehensions",
  "flake8-docstrings",
  "flake8-eradicate",
  "flake8-isort",
  "flake8-pytest-style",
  "flake8-spellcheck",
  "pep8-naming"
]
vim = ["black", "pynvim"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.coverage.run]
concurrency = ["multiprocessing"]
command_line = "-m pytest"
parallel = true
source = ["src"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
